12_quiz_bot/requirements.txt
aiogram==3.17.0
aiosqlite==0.21.0
python-decouple==3.8
